requirements.txt
Pillow
baidu-aip
facebook-wda